# docs/api/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/enrollments/{id}": {
      "get": {
        "operationId": "OperationalController_enrollment_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Gets an organization enrollment",
        "tags": [
          "Operational CRUD"
        ]
      },
      "patch": {
        "operationId": "OperationalController_updateEnrollment_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2000000000
                  },
                  "chargeOpenDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "chargeOpenTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "discountCents": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 2000000000,
                    "default": 0
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "ENDED",
                      "CANCELLED"
                    ]
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates an organization enrollment",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
...
    "/api/v1/workspace/public-enrollment-form": {
      "get": {
        "operationId": "PublicEnrollmentFormController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Gets the authenticated organization's public enrollment form",
        "tags": [
          "Public enrollment form"
        ]
      },
      "post": {
        "operationId": "PublicEnrollmentFormController_create_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates the organization's single public enrollment link",
        "tags": [
          "Public enrollment form"
        ]
      },
      "patch": {
        "operationId": "PublicEnrollmentFormController_update_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates public enrollment settings",
        "tags": [
          "Public enrollment form"
        ]
      }
    },
    "/api/v1/workspace/public-enrollment-form/rotate": {
      "post": {
        "operationId": "PublicEnrollmentFormController_rotate_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Rotates the public enrollment link and invalidates the previous one",
        "tags": [
          "Public enrollment form"
        ]
      }
    },
    "/api/v1/workspace/public-enrollment-form/submissions": {
      "get": {
        "operationId": "PublicEnrollmentFormController_submissions_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Lists public enrollment requests for the authenticated organization",
        "tags": [
          "Public enrollment form"
        ]
      }
    },
    "/api/v1/workspace/public-enrollment-form/submissions/{id}/approve": {
      "post": {
        "operationId": "PublicEnrollmentFormController_approve_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Approves a public enrollment request atomically",
        "tags": [
          "Public enrollment form"
        ]
      }
    },
    "/api/v1/workspace/public-enrollment-form/submissions/{id}": {
      "delete": {
        "operationId": "PublicEnrollmentFormController_reject_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Permanently deletes a pending public enrollment request",
        "tags": [
          "Public enrollment form"
        ]
      }
    },
...
    "/api/v1/public/enrollment/{token}/photo": {
      "post": {
        "operationId": "PublicEnrollmentController_uploadPhoto_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Uploads the required student photo for a public enrollment",
        "tags": [
          "Public enrollment"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mensaly API",
    "description": "Versioned backend API for Mensaly company owners and platform administrators",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mensaly_session"
      }
    },
...
  }
}